fix defect + create module
add new descriptions
